for RH: link /usr/lib/nagios to /usr/lib64/nagios
In our manifests, we commonly declare checks like:

  nagios::nrpe::service {
     'check_all_disks':
      check_command => '/usr/lib/nagios/plugins/check_disk -w \'10%\' -c \'5%\' -x cgroup -x none -x /dev -x /run'    ;
  }

This will create a file with the content of check_command in /etc/nagios/nrpe.d
but it will not create the check itself. The binary check_disk is packaged,
and in RedHat based distros, it appears in /usr/lib64/nagios/plugins. This is
still the case for Centos 7.

This change should mean that the checks we already have declared should work
for RH based systems.

Change-Id: Id5ea5797a519c9d20e6da9ef3236606b97798bd5